feat: work in progress - Fix LineCollection enumerator implementation - Remove AggressiveInlining on methods where it was regressing codegen quality and performance on reference benchmark (warpskimmer) - Add more extension methods because overload resolution does not infer implicit casting

+++ src/Impl/U8Conversions.cs
@@ -36,10 +36,10 @@ static class U8Conversions
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static ref T Offset<T>(this ref T value, nint offset)
+    internal static ref T Offset<T>(this ref T value, int offset)
         where T : unmanaged
     {
-        return ref Unsafe.Add(ref value, offset);
+        return ref Unsafe.Add(ref value, (nint)(uint)offset);
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
